Added a navigation controller to the main view controller.
Added a collection view to the main view controller set up the constraints so that
the collection view fills the screen.

Set the custom class for the view cell of the main view controller.

Set the reusable identifier of the view cell of the main view controller.

Set the app title in the main view controller.

Hard code the cell size in the collection view to 185x278 for now.