Added an outlet for the collection view.
Added the UICollectionViewDelegate and UICollectionViewDataSource protocols.

Added delegate and data source for the collection view.

Added the functions numberOfItemsInSection and cellForItemAt so that we conform to
the data source protocol.

Registered the movie collection xib file.

Added a handler for movie results that are received asynchronously.

Display movie posters in the collection view cells.