fix: Fix a crash triggered by state selection
When the user changes the mode that is displayed in this view controller
(i.e. changing from popular movies to favorites) the app could crash. The
reason is that the array backing the collection view is clear during a
mode change, but we fail notify the collection view. So the collection
view thinks that there are "n" number of items in the array when there
could be 0.

This change notifies the collection view when we change the data
source.