feat: Handle portrait orientation
This change will scale the movie posters when in portrait mode taking
into account the client specified number of posters in a row
(2 for now) and the horizontal gap between posters.

We delay the call to fetch the posters until we have received our first
rotation event. This is done so that we don't initialize the collection
view prematurely only to have to reinitialize the collection view after
the first rotation event occurs.